# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "CreateOrderRequest": {
        "properties": {
          "appointment_date": {
            "description": "预约日期",
            "examples": [
              "2026-05-01"
            ],
            "pattern": "^\\d{4}-\\d{2}-\\d{2}$",
            "title": "Appointment Date",
            "type": "string"
          },
          "appointment_time": {
            "description": "预约时间 HH:MM",
            "examples": [
              "09:30"
            ],
            "pattern": "^\\d{2}:\\d{2}$",
            "title": "Appointment Time",
            "type": "string"
          },
          "companion_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "可选：直接指派的陪诊师 ID；为空则进入大厅抢单",
            "title": "Companion Id"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "补充说明（病情、特殊需求）",
            "examples": [
              "需要陪同做核磁，行动不便"
            ],
            "title": "Description"
          },
          "hospital_id": {
            "description": "目标医院 ID",
            "format": "uuid",
            "title": "Hospital Id",
            "type": "string"
          },
          "service_type": {
            "description": "服务类型：full_accompany 全程陪诊 / half_accompany 半天陪诊 / errand 跑腿代办",
            "examples": [
              "full_accompany"
            ],
            "pattern": "^(full_accompany|half_accompany|errand)$",
            "title": "Service Type",
            "type": "string"
          }
        },
        "required": [
          "service_type",
          "hospital_id",
          "appointment_date",
          "appointment_time"
        ],
        "title": "CreateOrderRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OrderListResponse": {
        "properties": {
          "items": {
            "description": "当页订单",
            "items": {
              "$ref": "#/components/schemas/OrderResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "total": {
            "description": "总条数",
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total"
        ],
        "title": "OrderListResponse",
        "type": "object"
      },
      "OrderResponse": {
        "properties": {
          "appointment_date": {
            "description": "预约日期",
            "examples": [
              "2026-05-01"
            ],
            "title": "Appointment Date",
            "type": "string"
          },
          "appointment_time": {
            "description": "预约时间",
            "examples": [
              "09:30"
            ],
            "title": "Appointment Time",
            "type": "string"
          },
          "companion_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "陪诊师用户 ID（未接单时为空）",
            "title": "Companion Id"
          },
          "companion_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "陪诊师姓名（冗余）",
            "examples": [
              "张三"
            ],
            "title": "Companion Name"
          },
          "created_at": {
            "description": "创建时间",
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "患者补充说明",
            "title": "Description"
          },
          "expires_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "待支付订单的过期时间",
            "title": "Expires At"
          },
          "hospital_id": {
            "description": "医院 ID",
            "format": "uuid",
            "title": "Hospital Id",
            "type": "string"
          },
          "hospital_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "医院名称（冗余）",
            "examples": [
              "北京协和医院"
            ],
            "title": "Hospital Name"
          },
          "id": {
            "description": "订单 ID",
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "order_number": {
            "description": "订单号（业务可见）",
            "examples": [
              "YLA20260424100501"
            ],
            "title": "Order Number",
            "type": "string"
          },
          "patient_id": {
            "description": "患者用户 ID",
            "format": "uuid",
            "title": "Patient Id",
            "type": "string"
          },
          "patient_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "患者姓名（冗余）",
            "examples": [
              "小明"
            ],
            "title": "Patient Name"
          },
          "payment_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "支付状态",
            "examples": [
              "success"
            ],
            "title": "Payment Status"
          },
          "price": {
            "description": "订单金额（元，字符串形式避免浮点误差）",
            "examples": [
              "199.00"
            ],
            "title": "Price",
            "type": "number"
          },
          "service_type": {
            "description": "服务类型",
            "examples": [
              "full_accompany"
            ],
            "title": "Service Type",
            "type": "string"
          },
          "status": {
            "description": "订单状态",
            "examples": [
              "paid"
            ],
            "title": "Status",
            "type": "string"
          },
          "timeline": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/TimelineItem"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "订单时间轴",
            "title": "Timeline"
          },
          "timeline_index": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "当前时间轴索引",
            "title": "Timeline Index"
          },
          "updated_at": {
            "description": "最后更新时间",
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "order_number",
          "patient_id",
          "hospital_id",
          "service_type",
          "status",
          "appointment_date",
          "appointment_time",
          "price",
          "created_at",
          "updated_at"
        ],
        "title": "OrderResponse",
        "type": "object"
      },
...
      "TimelineItem": {
        "properties": {
          "time": {
            "description": "发生时间 ISO8601",
            "examples": [
              "2026-04-24T10:05:00+08:00"
            ],
            "title": "Time",
            "type": "string"
          },
          "title": {
            "description": "阶段名称",
            "examples": [
              "已接单"
            ],
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title",
          "time"
        ],
        "title": "TimelineItem",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {
      "email": "backend@yiluan.example.com",
      "name": "YiLuAn 后端团队"
    },
    "description": "\n# YiLuAn 一路安 后端 API\n\n「一路安」是一款连接患者与陪诊师的服务平台。本 API 文档涵盖：\n\n- **认证（auth）**：手机号 OTP 登录、微信登录、JWT 刷新\n- **用户与角色（users / patients / companions）**：账户、患者档案、陪诊师档案\n- **医院与订单（hospitals / orders）**：医院搜索、下单、状态流转\n- **支付与钱包（payment-callbacks / wallet）**：微信支付回调、钱包流水\n- **沟通（chats / notifications）**：订单内聊天、站内通知\n- **后台管理（admin / admin-companions）**：审核、强制状态、退款\n- **运维（health）**：liveness / readiness 探针\n\n## 鉴权\n\n绝大多数 `/api/v1/*` 接口需要在请求头携带：\n\n```\nAuthorization: Bearer <access_token>\n```\n\n`access_token` 由 `/api/v1/auth/verify-otp`、`/api/v1/auth/wechat-login`、\n`/api/v1/auth/refresh` 等端点签发。详见 [AUTHENTICATION.md](../../docs/api/AUTHENTICATION.md)。\n\n## 错误响应\n\n所有 4xx / 5xx 错误统一返回如下结构：\n\n```json\n{ \"detail\": \"human-readable message\" }\n```\n\n详见 [ERROR_HANDLING.md](../../docs/api/ERROR_HANDLING.md)。\n",
    "license": {
      "name": "Proprietary"
    },
    "title": "YiLuAn API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/admin/orders": {
      "get": {
        "description": "管理员查看所有订单列表，可按 `status` 过滤。仅 `admin` 角色可调用。",
        "operationId": "list_orders_admin_api_v1_admin_orders_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "后台：查询全部订单",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/orders/{order_id}/admin-refund": {
      "post": {
        "description": "以 `refund_ratio`（0~1）按订单金额按比例退款，1.0 表示全额退。",
        "operationId": "admin_refund_order_api_v1_admin_orders__order_id__admin_refund_post",
        "parameters": [
          {
            "in": "path",
            "name": "order_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Order Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "refund_ratio",
            "required": false,
            "schema": {
              "default": 1.0,
              "maximum": 1.0,
              "minimum": 0.0,
              "title": "Refund Ratio",
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "后台：管理员退款",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/orders/{order_id}/force-status": {
      "post": {
        "description": "管理员手动将订单跳转到指定状态，**仅用于运营干预**，不走业务状态机。",
        "operationId": "force_order_status_api_v1_admin_orders__order_id__force_status_post",
        "parameters": [
          {
            "in": "path",
            "name": "order_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Order Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "target_status",
            "required": true,
            "schema": {
              "title": "Target Status",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "后台：强制修改订单状态",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/users": {
      "get": {
        "description": "分页查看所有用户（含已停用）。",
        "operationId": "list_users_admin_api_v1_admin_users_get",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "后台：用户列表",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/users/{user_id}/disable": {
      "post": {
        "description": "将指定用户账号设为 `is_active=False`，用于风控处置。",
        "operationId": "disable_user_api_v1_admin_users__user_id__disable_post",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "User Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "后台：停用用户",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/users/{user_id}/enable": {
      "post": {
        "description": "重新启用被停用的账号。",
        "operationId": "enable_user_api_v1_admin_users__user_id__enable_post",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "User Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "后台：启用用户",
        "tags": [
          "admin"
        ]
      }
    },
...
    "/api/v1/orders": {
      "get": {
        "description": "分页查询当前用户参与的订单（患者视角看自己创建的，陪诊师视角看自己接的）。",
        "operationId": "list_orders_api_v1_orders_get",
        "parameters": [
          {
            "description": "订单状态过滤",
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "订单状态过滤",
              "title": "Status"
            }
          },
          {
            "description": "预约日期 YYYY-MM-DD",
            "in": "query",
            "name": "date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^\\d{4}-\\d{2}-\\d{2}$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "预约日期 YYYY-MM-DD",
              "title": "Date"
            }
          },
          {
            "description": "按城市过滤",
            "in": "query",
            "name": "city",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "按城市过滤",
              "title": "City"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "Not authenticated"
                }
              }
            },
            "description": "未鉴权或令牌无效"
          },
          "422": {
            "content": {
              "application/json": {
                "example": {
                  "detail": [
                    {
                      "loc": [
                        "body",
                        "phone"
                      ],
                      "msg": "Invalid phone number format",
                      "type": "value_error"
                    }
                  ]
                }
              }
            },
            "description": "校验失败（FastAPI 标准）"
          },
          "500": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "Internal server error"
                }
              }
            },
            "description": "服务器内部错误"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "获取我的订单列表",
        "tags": [
          "orders"
        ]
      },
      "post": {
        "description": "患者发起一笔陪诊服务订单。需指定服务类型、医院、就诊日期与时间。可选 `companion_id` 直接指派，否则进入大厅由陪诊师抢单。\n\n新订单状态为 `pending_payment`，**必须在 30 分钟内完成支付**，否则会被定时任务自动取消。",
        "operationId": "create_order_api_v1_orders_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrderRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "Invalid request payload"
                }
              }
            },
            "description": "请求参数错误"
          },
          "401": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "Not authenticated"
                }
              }
            },
            "description": "未鉴权或令牌无效"
          },
          "422": {
            "content": {
              "application/json": {
                "example": {
                  "detail": [
                    {
                      "loc": [
                        "body",
                        "phone"
                      ],
                      "msg": "Invalid phone number format",
                      "type": "value_error"
                    }
                  ]
                }
              }
            },
            "description": "校验失败（FastAPI 标准）"
          },
          "500": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "Internal server error"
                }
              }
            },
            "description": "服务器内部错误"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "患者创建订单",
        "tags": [
          "orders"
        ]
      }
    },
...
  },
...
}